v3.0.0

+++ src/main/java/ru/edu/project/lab/service/ComposerService.java
@@ -2,11 +2,12 @@ package ru.edu.project.lab.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.edu.project.lab.baseresponse.ErrorType;
 import ru.edu.project.lab.entities.Composer;
 import ru.edu.project.lab.Repository.ComposerRepository;
 import ru.edu.project.lab.dto.ComposerDto;
+import ru.edu.project.lab.exception.PenzGtuException;
 import ru.edu.project.lab.service.mapper.ComposerMapper;
-
 import java.util.List;
 
 @Service
@@ -15,34 +16,39 @@ public class ComposerService {
     private final ComposerRepository composerRepository;
     private final ComposerMapper composerMapper;
 
-    public List<ComposerDto> findAllAuthors() {
+
+    public List<ComposerDto> findAllComposers() {
 
         return composerMapper.toListDto(composerRepository.findAll());
     }
 
-    public ComposerDto findAuthorById(long id) {
+    public ComposerDto findComposerById(long id) {
         Composer composer = composerRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Author not found"));
+                .orElseThrow(()-> new PenzGtuException(ErrorType.NOT_FOUND,"Error finding composer by id"));
         return composerMapper.toDto(composer);
     }
 
-    public void saveAuthor(ComposerDto composerDto) {
+    public void saveComposer(ComposerDto composerDto) {
         Composer composer = composerMapper.toEntity(composerDto);
 
         composerRepository.save(composer);
 
     }
 
-    public void updateAuthor(ComposerDto newComposer) {
+    public void updateComposer(ComposerDto newComposer) {
         Composer oldComposer = composerRepository.findById(newComposer.getId())
-                .orElseThrow(()-> new RuntimeException("Author not found"));
-
+                .orElseThrow(()-> new PenzGtuException(ErrorType.NOT_FOUND,"Error finding composer by id"));
         oldComposer.setName(newComposer.getName());
 
         composerRepository.save(oldComposer);
     }
 
-    public void deleteAuthorById(long id) {
+    public void deleteComposerById(long id) {
+        if (!composerRepository.existsById(id)) {
+            throw new PenzGtuException(ErrorType.NOT_FOUND,"Error finding composer by id");
+        }
+
         composerRepository.deleteById(id);
+
     }
 }
